requirements.txt
google-generativeai
supabase
python-dotenv
pdfplumber
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client